app_utils/WaveNumberOCR/requirements.txt
pillow
mss
easyocr
